services/ask-bud/requirements.txt
# Service/Clients
dapr >= 1.14.0
budmicroframe @ git+https://github.com/BudEcosystem/bud-microframe.git

# API
fastapi >= 0.111.1
uvicorn >= 0.30.3
pydantic >= 2.8.2
pydantic-settings >= 2.3.4

# Utils
structlog >= 24.4.0
cachetools >= 5.4.0

# Database
SQLAlchemy>=2.0.36
sqlalchemy-utils>=0.41.2
psycopg
psycopg2-binary>=2.9.10
alembic>=1.13.3
alembic-postgresql-enum>=1.3.0

# Agent
openai-agents